DI 적용: ChatServerClient
